requirement.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.19
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
dlx==1.0.4
docplex==2.14.186
entrypoints==0.3
fastdtw==0.3.4
fastjsonschema==2.14.4
h5py==2.10.0
idna==2.9
inflection==0.4.0
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-polyfield==5.9
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.3.0
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.6
nest-asyncio==1.3.3
networkx==2.4
notebook==6.0.3
ntlm-auth==1.4.0
numpy==1.18.4
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
pybind11==2.5.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-constraint==1.4.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.1
qiskit==0.19.2
qiskit-aer==0.5.1
qiskit-aqua==0.7.1
qiskit-ibmq-provider==0.7.1
qiskit-ignis==0.3.0
qiskit-terra==0.14.1
qtconsole==4.7.4
QtPy==1.9.0
Quandl==3.5.0
requests==2.23.0
requests-ntlm==1.1.0
retworkx==0.3.4
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
sympy==1.6
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
websockets==7.0
widgetsnbextension==3.5.1
